Move tabbar container out of containers, as its only bar, no container
